Don't make modes and binary field null for renames
We don't get this info from Git for re-names. It felt funky to say that the modes were null when the files actually have modes, we just don't know what they are. Plus, we use a null file mode to indicate that the file does not exist for created/deleted files.

Co-authored-by: Michelle Tilley <michelle@michelletilley.net>